+++ src/main/java/com/warxim/petep/extension/internal/history/gui/view/HistoryViewController.java
@@ -123,6 +123,7 @@ public class HistoryViewController implements Initializable {
         // Refresh items
         view.setFilter(view.getFilter());
         initContextMenu();
+        metadataPane.setVisible(false);
     }
 
     /**